+++ register/models.py
@@ -8,4 +8,4 @@ class user_data(models.Model):
 	name= models.CharField(max_length= 120,blank= True ,null = True)
 	email= models.TextField(blank= True, null= True)
 	city= models.CharField(max_length=120, blank=True, null=True)
-	mobile=models.PositiveSmallIntegerField(primary_key=True,default=0)
+	mobile=models.IntegerField(primary_key=True,default=0)
